Renamed Scope to InterceptorGroupTransaction. Moved plugin annotation types to plugin.annotation package.
