+++ bootstrap/src/main/java/com/navercorp/pinpoint/bootstrap/interceptor/group/InterceptorGroup.java
@@ -21,5 +21,5 @@ package com.navercorp.pinpoint.bootstrap.interceptor.group;
  */
 public interface InterceptorGroup {
     public String getName();
-    public InterceptorGroupTransaction getCurrentStack();
+    public InterceptorGroupTransaction getCurrentTransaction();
 }
